Fetches from TemplateListPage.tsx and CoursesListPage.tsx moved to separate files. LocalHostAddresses.ts moved to api different directory
