--- frontend/src/LocalHostAddresses.ts
@@ -1,7 +1,0 @@
-enum Addresses{
-    AUTH = "https://localhost:5006",
-    COURSESTREAMS = "https://localhost:5237",
-    COURSETEMPLATES = "https://localhost:5204",
-    MATERIALS = "https://localhost:5116"
-}
-export {Addresses};
