+++ frontend/src/api/coursesListService.ts
@@ -1,0 +1,7 @@
+import { Addresses } from "./LocalHostAddresses";
+
+export const getCoursesList = async () => {
+    const response = await fetch(`${Addresses.COURSESTREAMS}/courseStreams/list`);
+    const data = await response.json();
+    return data.data; // Assuming the API returns an array of courses
+};
