+++ frontend/src/api/templateService.ts
@@ -1,0 +1,7 @@
+import { Addresses } from "./LocalHostAddresses";
+
+export const getCourseTemplates = async () => {
+    const response = await fetch(`${Addresses.COURSETEMPLATES}/courseTemplates/list`);
+    const data = await response.json();
+    return data.data; // Assuming the API returns an array of course templates
+};
